Updated now.html IMPORTANT
The later and done are not being focused on at the moment and so they are incomplete and do not look like "now.html". "Now.html" is the main testing page.
This now page will not work when pulling items from your database on your local machine. I have added myself as a user and am testing pulling the data with my own user when logged in.